created new class CryptoUtilsTest and moved tests.
the last CryptoUtils tests were in the placeholder file.
